The addition of new WordPress routes is being handled by the WordPressRouter. Extracted the WordPress specific logic for the Routes to The WordPressRoute class and added a new WPRoute facade which points to the WordPressRouter.
